Update Java7
For example, if the string  s='abcac' and , the substring we consider is abcacabcac, the first 10 characters of her infinite string. There are  4occurrences of a in the substring.
1. Function Description
Complete the repeatedString function in the editor below. It should return an integer representing the number of occurrences of a in the prefix of length  in the infinitely repeating string.

repeatedString has the following parameter(s):
s: a string to repeat
n: the number of characters to consider

2. Input Format
The first line contains a single string, s
The second line contains an integer, n

3. Constraints
1 <= s <= 100
1 <= n <= 10^12
For 25% of the test cases, n <= 10^6

4. Output Format
- Sample Input
aba
10
- Sample Output
7
- Sample Input
a
1000000000000
- Sample Output
1000000000000